all 4 module
2 modules(Vol and Month) search and  4 modules(Vol,Month,FieldOpr and Header) validation

+++ ihsprod-header/entities/ihsprod-header.entity.ts
@@ -55,7 +55,7 @@ export class IHSProdHeaderEntity extends CustomBaseEntity {
   @Column({ type: 'varchar', length: 80, nullable: true })
   lease_name: string;
 
-  @Column({ type: 'int', width: 10, nullable: true })
+  @Column({ type: 'int8', width: 10, nullable: true })
   lease_number: number;
 
   @Column({ type: 'varchar', length: 10, nullable: true })
